347920 [gerrit] Open commit viewer from task editor
This commit adds a link to the patch set that allows to open a commit
in the commit viewer by clicking on the SHA1. If the commit is not
available in the repository yet, it is fetched.

Some methods and classes are moved out of PatchSetSection because the
class would grow way too large.

Bug: 347920
Change-Id: I5c73c0bcc47a52b7efb7964028598467d43f2af6