Suppress parse exceptions deserializing event payloads
Older events may include payloads with different field
names or value types and parse exceptions that occur
currently cause the entire list parsing to fail.

By catching the parse exception on payload deserialization
the event can still be returned with core fields but just
a missing payload.

This will allow returned pages with a mix of legacy and new
event payloads to return successfully with proper payloads
for new events and null payloads for legacy events.

Change-Id: Iac866dd2d7a68ebbcb4ea7bec0c567a37e560ba0