bug 380843: index commit messages of changesets
Added an initial implementation of an index for the scm repositories.
The commit message, the revision and repository url are indexed, which
allows to search for all commits for a task in specific repositories.
Currently, the index is rebuild at the startup, which limits its use
only for smaller git projects. This should be fixed in the near future,
when a notification of repository changes is implemented, as well as
a mechanism, which determines if the repository has changed while
eclipse was not running.

Change-Id: If7aebd06dd9ca19090b2af3b5f447fe8454df7d0