410655: redistribute JDT annotations to make bundle available for implicit target builds
Since the annotations are not shipped as of older Eclipse SDKs make them
available from the commons nightly site so other builds and developers
can consume the bundle easily. 

The change also adds an explicit target for consuming the annotations
bundle from the upstream repository. An implicit target wouldn't work
well since the bundle is published on the 3.8 SDK site which has other
undesired bundles.

Change-Id: I5d59eaa7e49652d14e77f136fbd13d53be1aec85
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=410655