430926: Sort projects list alphabetically when importing from Gerrit
Order of the projects list returned by GerritClient#getVisibleProjects
is more or less undetermined. The order of projects is one of the
entries in the intermediate HashMap used to parse Gerrit server
response. With this change projects list returned by
GerritClient#getVisibleProjects is explicitly sorted by project name.
Ordering is ascending, case insensitive with nulls last. 

This change will indirectly make projects list alphabetically sorted on
the project selection page of the "Projects from Git" wizard with Gerrit
as the source repository.

Bug: 430926
Change-Id: Ia1e5aa3fe833013cdd4446eae8e42ce758b5c2f5
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=430926
Signed-off-by: Vadim Dmitriev <dmgloss@mail.ru>