431157: Accept gerrit URLs which include hash character
This allows copy-paste of gerrit review URLs from the browser to the
task repository manager, even if the URL points to a specific review,
filter, dashboard, or similar.

This change introduces a restricted access warning for calling
TaskRepositoryManager.stripSlashes(). This was done on purpose, as the
super implementation of getRepositoryUrl also cleans up using
stripSlashes.

Change-Id: I5e39137f0706d5e106e4ecb0da14f45fe70cbecc
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=431157
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>