Added `context` field to the CommitStatus class
Relevant GitHub API docs: https://developer.github.com/v3/repos/statuses/
This field is used to indicate the service that provided the commit status in
question. Its value is an (apparently arbitrary) string with a default value
of "default". Conventionally, the value should be of the form
"generic-category/specific-service", as in "continuous-integration/jenkins"
or "security/brakeman".

Bug: 461351
Change-Id: Ifef4e369c4d602ac6a5d1f12c399d059052b8abc
Signed-off-by: Chris Rebert <eclipse@rebertia.com>