463996: provide Oomph setup files for Mylyn
* Provides setup files for the following projects: Commons, Docs,
Context, Tasks, Bugzilla REST Connector.
* A work-around is required for the JDT annotations dependency: Since
context depends on JDT which includes org.eclipse.jdt.annotation version
2.0, bundles using @NonNull annotations get compile errors. The reason
is that Mylyn needs the 1.1 version of the bundle since the 2.0 version
requires Java 8. To force resolving against the 1.1 version it is
imported into the workspace from the eclipse.jdt.core repository when
Mylyn Context is provisioned. 

Change-Id: I8cebeca0a7f62721a089344ef4686fd94d3bd59b
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=463996