Resolve compilation problem on Oxygen
org.eclipse.mylyn.internal.github.ui.gist makes use of many
internal classes from org.eclipse.mylyn.tasks.ui. With version
3.23.0 of Mylyn Tasks, one of these classes changes in incompatible
ways, so this part of egit-github no longer compiles.

As a quick fix, take an slightly older version of this class from
Mylyn Tasks, rename it, and minimally adapt it to get it to compile.
Then use that copy. This will work with Mylyn Tasks versions 3.7 up
to at least 3.23.0.

This is a hack. For one, the class in question is a label provider,
and the using class already fiddles with column indices to adjust
that label provider to its particular use. So a more complete fix
would be to write our own, dedicated label provider. And of course
there's no guarantee that egit-github won't break again when other
Mylyn-internal classes change.

Bug: 516148
Change-Id: I03896c43f6876821ca5cc1f95d8fcebcb05e9812
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>