Support diff/patch generation via the Github API
The Github API can generate .diff and .patch files server-side for individual
commits and ranges of commits. This feature is very useful, as it permits one
to fetch the contents of file changes without needing to download an entire
repository. See: https://developer.github.com/v3/repos/commits/.

This change adds support for these features to the CommitService.

Bug: 529255
Change-Id: I9c700f0030578299283021b8f645d052bccca3c6
Signed-off-by: Christian Barcenas <christian@cbarcenas.com>