533119: First line of commit template ends with superfluous blank
The commit message template first line ends with a superfluous blank,
which is taken over as such when a commit message is created in egit.
Let's remove it from the template default value.

Also fix an off-by-one error in the commit template parsing. With the
new fixed template, it would always lead to removing the trailing curly
brace, and the substring method is already excluding the end index by
itself.

Change-Id: If0f138b10dafe1d0b3bd09c30c3bc10eec5eeb65
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Bug:533119