Fix Repository.getWatchers(); add Repository.getStars()
Per [1] it appears that the JSON returned by Github changed,
apparently already in 2012. What Github shows today as the number of
"watchers" is stored in field "subscribers_count". Field "watchers"
actually is a copy of "stargazers_count" (as is "watchers_count").

So add fields "stargazersCount" and "subscribersCount", but keep the
legacy "watchers" field to be able to deal with old serialized (via
Gson or via Java serialization) objects.

Add tests, including a new test for Gson de-serialization. The test
data in test_repo.json is anonymized actual output for an existing
repository obtained from the Github API as of 2018-05-08.

[1] https://developer.github.com/changes/2012-09-05-watcher-api/

Bug: 534440
Change-Id: I1faba649986ab6ed9e367bfefdb0bbe2fbaf7071
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>