535784: HtmlCleaner preserves hyperlinks
Changed HtmlCleaner to preserve HTML anchors when a non-empty href
attribute is present.  This affects content where that contains anchors
containing nothing but whitespace within the <a> tags.

Previously such tags were removed by the cleaner, with the assumption
being that empty anchor tags were not visible to the user.  The
motivation for this change is to support content where CSS or other
styling makes empty anchor tags significant.

Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=535784
Change-Id: Ia804de343131c7da8f2e9484b6ec5b3cdbad61af
Signed-off-by: Leo Dos Santos <leo.dos.santos@tasktop.com>