Remove outdated github launch file
This launch file is completely outdated and doesn't work at all. Even
changes of the last years were basically always just because of mass
updates. It references workspace projects which are not of the eclipse
foundation and it lives in the .settings folder, which is not a good
place for launch configs.

Since egit developers use local launch configs, the same approach seems
reasonable for egit-github, so delete this one.

Change-Id: I36b6359d5398eb9a6d7a6cda54878e843205f74d
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>