[releng] Store all project properties explicitly
We use project specific properties in many projects. However, the files
containing them are old, therefore they do not contain all the keys and
values that have been added in newer Eclipse releases.

This change makes all those missing settings explicit. All project
properties have been opened in the UI and the "Apply" button has been
used without changing any controls. We should therefore see only
additions in the diff, no changes in values.

The main intention of this change is to reduce the noise for actual
changes of settings which are intended to be applied afterwards.

The removal of JDT settings in org.eclipse.mylyn.github.doc is
intentional. That is a PDE project without Java nature.

Change-Id: I66b75fc8a64660615a0d2a8601f0b0cf3de82993
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>