540235: ExtendedQuoteBlock supports nested blocks over multiple lines
* Override the three functions necessary for nested block supported as
specified in the Block javadocs: beginNesting(), findClosedOffset(), and
canResume(). This way we can leverage the logic within
AbstractMarkupLanguage that manages the state of nested blocks.

* Further in order for AbstractMarkupLanguage to supported
ExtendedQuoteBlock and its nested blocks it needs to know what the
offset is for the current line being processed. Currently the abstract
function in AbstractConfluenceDelimitedBlock assumes the whole line
being processed and has a void return. This is not the case so we should
return an offset when calling handleBlockContent().

* Added tests for the transformation of Confluence markup to HTML

* Added tests for the transformation of HTML document building blocks to
Confluence markup

https://bugs.eclipse.org/bugs/show_bug.cgi?id=540235
Change-Id: Ic11a7605fc0bbca3c97a6e432a4036834faafd33
Signed-off-by: Ryan Nosworthy <ryan.nosworthy@tasktop.com>