542496: Add support for Tables to the Creole DocumentBuilder
Note that the spec for Tables in Creole found at
http://www.wikicreole.org/wiki/Tables indicates 'table rows ending with
| are optional.'

I elected to end table rows with '|' in the Creole Document Builder
because the implementation of CreoleLanguage currently found in wikitext
only parses table rows which include the '|' suffix. Consistency over
completeness - though I think a future feature request should address
amending the CreoleLanguage to recognize a line such as '|Cell A|Cell B'
as a valid table row.

Change-Id: I11c0c7cfbc4ea632d9b847261e3d8f25754d4147
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=542496
Signed-off-by: Kevin de Vlaming <kevin.devlaming@tasktop.com>