542496: Fix escape behavior in Creole Document Builder
- The previous approach of replacing the tilde literal with its entity
reference when writing to Creole was incorrect, as it would not be
converted back to a literal when parsed via Creole Language
- A better approach is to escape any literal tildes by adding a
preceding tilde. Ie., '~' will be written to Creole as '~~'.
- CreoleLanguage is capable of detecting the first of these two tildes
as an escape then reading the second as a literal, thus ensuring that
during the round trip the tilde is kept intact
- Also added some additional special characters that should be escaped
with a tilde

Change-Id: If0ed1c10ed6ff961f1f15a02ec36fa693b2c6ee3
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=542496
Signed-off-by: Kevin de Vlaming <kevin.devlaming@tasktop.com>