Bug 506804 - fixed wrong sort order of substring proposals
Enable needsSortingAfterFiltering for FocusedJavaAllProposalComputer,
same as in JDT.

When content assist is triggered by '.' on a map, "compute" is ordered
before "put" because they have the same relevance and are then ordered
alphabetically. Once you type "put", the relevance of "compute"
decreases as it's now a substring match.
Therefore the proposals needs to be re-sorted.

Change-Id: I6f8b615b2112cf04e394b19f315e9d151dcb1f27
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=506804
Signed-off-by: visjho <julian.honnen@vector.com>