Support token-based authentication
Change the repository wizard pages such that the user can enter a
token instead of username/password. The token is stored in the normal
password property; we add an extra property that records that it
is a token. Then use token-based authentication instead of basic
authentication for all http/https requests.

In files touched also modernize the code a little. Minimum version of
Mylyn supported in 3.20, so several internal accesses can be replaced
by proper API.

Bug: 567557
Change-Id: I566439e44a9000f313e2801890b1e3b5c4c1903f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>