[token auth] Fetch user for token if needed
There are some places in egit-github that assume that a user name is
available. With token auth, the user needs not be given. Make the
GitHubClient fetch it automatically if token auth is used.

This should work fine for PRs and for issues. For gists, still require
the user name be given in the task repository configuration. The code
assumes it can get the user name from there in several places where
there is no GitHubClient easily accessible.

Bug: 567557
Change-Id: I5baab8c21759711fcf46897530c81e7b38f023be
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>