571478: Prevent Confluence Wikitext from outputting multiple \\ per line
Multiple \\ without spaces in a line do not render as line breaks. This
reverts the change made to isBlockTerminatedByNewlines for bug 565028,
to how it was initially added for bug 485456. The portion of the change
being reverted was mentioned as a refactoring in the commit where it was
introduced and should not affect the fix for that defect (the automated
tests validate this and manual testing was performed). This patch also
includes an additional test for the escaped content fix from bug 565028.

Change-Id: Id19e07a7ceb6202fb96bbb7d9f97f3281c838021
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571478
Signed-off-by: Myles Feichtinger <myles.feichtinger@tasktop.com>