580782: Update Mylyn Docs POMs to use CI Friendly versions
Update the Mylyn Docs POMs to use the extended (Tycho) CI Friendly
properties for project versions.

* The releaseVersion property in the root POM is set to the current
wikitext development version (3.0.44).
* The qualifier property in the root POM is set to '-SNAPSHOT' by
default. It will be overridden to a Tycho-generated timestamp during a
release build.
* The target-platform, epub and docs projects maintain their own
independent versions that will remain as-is. But they will use the
qualifier property to enable release builds.

The flatten maven plugin has been integrated with the root POM to
ensure valid POMs can be generated for deployment.

Change-Id: I4b6ddbcae7557131f180ba78da3d998fb9d675a2
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580782