580782: Fix mylyn docs maven deployment
Introduce the nexus staging maven plugin to replace the maven deploy
plugin. This new plugin will eventually allow automated promotion to a
release instead of executing manual promotion steps inside the Nexus
UI. Further details in
https://central.sonatype.org/publish/publish-maven/. For now though,
set the autoReleaseAfterClose = false for testing/troubleshooting.

Only a subset of the maven modules should be deployed to Nexus and
eventually maven central. Thus, the plugin's skipNexusStagingDeployMojo
is set to true for the modules that should be excluded.

Unfortunately there is a limitation with this plugin where the
deployment will be completely skipped if the last module in the reactor
has skipNexusStagingDeployMojo = true. That is the case here because
the docs-site artifact (eclipse p2 repository) should not be deployed
but it needs to depend on everything else.

As a workaround, introduce a dummy aggregate module that depends on
the docs-site artifact to ensure the aggregate module runs last (with
deployment enabled). This follows a workaround documented on
https://issues.sonatype.org/browse/NEXUS-9138.

Change-Id: I848104b167103c2ad99b2760fff12d9a00840d6b
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580782