580782: Enforce Java 11 compatibility for compiler.
Java 9 and above expose a '-release' parameter that tells the compiler
to enforce API and binary (classfile) compatibility with the given
version. This is a stronger guard than the '-source' and '-target'
options. If any code uses language features or API's that don't
exist in the given version, the compiler will emit an error and fail
hard.

For Mylyn Docs, enforce Java 11 compatibility.

Change-Id: I51258aee9a1d0ad7ea39c3cc32bfbd44d99e0493
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580782