580782: SCM plugin connects to mylyn docs git over SSH
Configure the maven SCM plugin to use SSH for connecting to the mylyn
docs git repository. Connecting via https is unsupported within the
context of the jenkins job. SSH credentials are already specified
in the Jenkins configuration.

The proper SSH-based git url is defined already in the POM's
SCM.developerConnection property so reference that from the scm plugin
configuration.

Change-Id: I0e3de630d24eddf1c7f95be2331283a9e164ff7d
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=580782