Bundle plugins do not generate imports for java.* packages
Create wikitext and epub OSGI bundles that do not add import-package
declarations on java.* packages. Excluding java.* packages maintains
compatibility with pre-R7 OSGI frameworks.

Declaring explicit bundle imports on java.* packages was introduced
with Java 11 as a way for OSGI frameworks to validate the availability
of core java packages in the modularized JRE that started with Java 9.
However it's unnecessary to declare these bundle imports as documented
in the OSGI Core R7 specification
https://docs.osgi.org/specification/osgi.core/7.0.0/framework.module.html#framework.module-execution.environment.