Introduce Mylyn Docs Jenkins release pipeline job
A new jenkinsfile defines the build and release steps that the
associating Jenkins job
(https://ci.eclipse.org/mylyn/view/Docs/job/github-mylyn-docs-release/)
will perform. This pipeline is configured to run only against the
master branch (trunk).

The pipeline is split into the following stages:

* Initialize PGP: initialize the PGP keys that will be used for
jar signing.

* Build and Deploy Maven Artifacts: builds each of the mylyn docs
artifacts (jars, p2 repos, etc.) and run all tests. On a successful
build, a subset of the wikitext artifacts will be deployed to Nexus.

* Deploy Update Site: copies the docs-site p2 repository from the
previous build step to the download.eclipse.org file share.

* Tag Release: tags the commit associated with the mylyn docs
release that was just created.

* Bump Development Version: increments the patch version component
of the wikitext version (e.g. 3.0.45 -> 3.0.46).

The latter 3 stages can be restarted for a given build using the
built-in functionality of the Jenkins Pipeline.

By default, the build steps will execute in dry-run mode
(PERFORM_RELEASE = false). In this mode, the deployed maven artifacts
will not be released, the docs-site p2 repository will not be copied
to download.eclipse.org, and tags/commits will not be pushed to the
git repository.