Release pipeline releases nexus staging repository automatically
The nexus-staging-maven-plugin is used to deploy wikitext artifacts
to a staging repository in Nexus. The plugin has the ability to
release the staging repository automatically if the associating
autoReleaseAfterClose parameter = true. This means it's unnecessary
to manually release the staging repository using the Nexus
Repository Manager.

Unfortunately there was a flaw in the root POM that always set
the autoReleaseAfterClose parameter to false, even if the user property
-DautoReleaseAfterClose=true was passed on the command-line (as it
would if the PERFORM_RELEASE build parameter was true). The nexus
staging plugin gives precedence to the configuration supplied in the
POM.

The plugin configuration in the POM has been updated so that it is
now possible to override the autoReleaseAfterClose parameter.