Remove bugzilla from Mylyn IDE feature
It used to be that the Mylyn IDE feature didn't really pull in
any connectors, but just IDE features. However there was a log
message about bugzilla as repored in #168

The current fix for #168 pulls in more of bugzilla when the ide_feature
is installed.

Therefore this change is to not pull in bugzilla at all when
installing IDE feature.

Fixes https://github.com/eclipse-mylyn/org.eclipse.mylyn/issues/168